fix color override issue
- don't force color if an actor is related to the unit

+++ xlive/Blam/Engine/Simulation/game_interface/simulation_game_objects.h
@@ -12,10 +12,10 @@ struct s_simulation_object_state_data
 	real32 scale;
 	real_vector3d translational_velocity;
 	real_vector3d angular_velocity;
-	int body_vitality;
+	real32 body_vitality;
 	char body_stun_ticks_is_zero;
 	char gap45[3];
-	int shield_vitality;
+	real32 shield_vitality;
 	char shield_stun_ticks_is_zero;
 	char dead;
 	char gap4D[2];
